Move tabs inside config folder under tabs.
