+++ packages/extension/src/view/devtools/pages/privateAdvertising/protectedAudience/prebid/tabs/config/components/consentManagement.tsx
